refactor(newrelic): rename sub packages
